UI updates and validation
Some validation added to ensure a user enters a string which could potentially be a URL (ensuring it is not blank and contains at least a '.').

Added some UI errors to display if the above is not shown (disabling of button, red message and outline).

Added safeguard in the routing to redirect if user bypasses the above and still submits something invalid.