Expiry Functionality
Beginning to add functionality which will allow a user to select when their link will expire (including UI element). Currently 1 minute, 1 hour, 1 day and 'never'.

As such, slight change to the active_tokens dict. The value is now a list containing the original URL at [0] and the timestamp of expiry at [1].

Comments and  references the dictionary have also been updated (to ensure we access the element in the list). Testing which used the dictionary will be updated in course.

A further method will be added which will be periodically run to ensure clean up is done if datetime.now() > expiry time (active_tokens[token_reference][1]).