Clean-up Function
Added a function which iterates over a copy of the active_tokens dictionary. If the time held in the [1] element (indicating the time of expiry) has passed, compared with a datetime taken during the method call, the key is removed from the dictionary.

Also a slight update to some comments to consistently describe the new data structure for active_tokens.